Added a screen Select feature to the game
The ability to select screens has been added to a new method in the game class. the organisation of the resources has been changed to be similar to the organisation of the classes. An options menu has been added with a tempary background and menu button in the main menu some code has also been tidied up and commented